+++ apps/monochromifier/src/utils/version-constants.js
@@ -1,3 +1,3 @@
 // This file is auto-generated during build - do not edit manually
-export const APP_VERSION = '0.6.0';
-export const BUILD_TIME = '2026-03-19T17:46:04.284Z';
+export const APP_VERSION = '0.7.0';
+export const BUILD_TIME = '2026-03-21T03:50:47.924Z';